+++ modular-service-admin-cli/api_module.json
@@ -1,5 +1,5 @@
 {
     "module_name": "modular_service",
-    "cli_path": "/modular_service_admin_cli/group",
+    "cli_path": "/group",
     "mount_point": "/modular_service"
 }
