+++ modular-service-admin-cli/group/region.py
@@ -1,10 +1,9 @@
 import click
 
-from modular_service_admin_cli.group import cli_response, ViewCommand
-from modular_service_admin_cli.service.constants import (PARAM_NAME,
-                                                         PARAM_PERMISSIONS,
-                                                         PARAM_ID,
-                                                         CLOUD_PROVIDERS)
+from group import cli_response, ViewCommand
+from service.constants import (
+    PARAM_NAME, PARAM_PERMISSIONS, PARAM_ID, CLOUD_PROVIDERS
+)
 
 
 @click.group(name='region')
@@ -20,7 +19,7 @@ def describe(maestro_name=None):
     """
     Describes Region.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().region_get(maestro_name=maestro_name)
 
 
@@ -38,7 +37,7 @@ def activate(maestro_name, native_name, cloud, region_id=None):
     """
     Activates Region.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().region_post(
         maestro_name=maestro_name, native_name=native_name, cloud=cloud,
         region_id=region_id)
@@ -52,5 +51,5 @@ def delete(maestro_name=None):
     """
     Deletes Region.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().region_delete(maestro_name=maestro_name)
