+++ modular-service-admin-cli/group/role.py
@@ -1,10 +1,8 @@
 from datetime import datetime
 
 import click
-from modular_service_admin_cli.group import cli_response, cast_to_list, ViewCommand
-from modular_service_admin_cli.service.constants import (PARAM_NAME,
-                                                         PARAM_POLICIES,
-                                                         PARAM_EXPIRATION)
+from group import cli_response, cast_to_list, ViewCommand
+from service.constants import (PARAM_NAME, PARAM_POLICIES, PARAM_EXPIRATION)
 
 
 @click.group(name='role')
@@ -19,7 +17,7 @@ def describe(name=None):
     """
     Describes roles.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().role_get(role_name=name)
 
 
@@ -35,7 +33,7 @@ def add(name, policies, expiration):
     """
     Creates the Role entity with the given name
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
 
     policies = cast_to_list(policies)
     if expiration:
@@ -63,7 +61,7 @@ def update(name, attach_policy, detach_policy, expiration):
     """
     Updates role configuration.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
 
     if not attach_policy and not detach_policy:
         return {'message': 'At least one of the following arguments must be '
@@ -93,5 +91,5 @@ def delete(name):
     """
     Deletes role.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().role_delete(role_name=name)
