+++ modular-service-admin-cli/group/tenant.py
@@ -1,10 +1,9 @@
 import click
-from modular_service_admin_cli.group import cli_response, ViewCommand
-from modular_service_admin_cli.group.tenant_region import region
-from modular_service_admin_cli.service.constants import (PARAM_NAME,
-                                                         PARAM_PERMISSIONS,
-                                                         PARAM_ID,
-                                                         CLOUD_PROVIDERS)
+from group import cli_response, ViewCommand
+from group.tenant_region import region
+from service.constants import (
+    PARAM_NAME, PARAM_PERMISSIONS, PARAM_ID, CLOUD_PROVIDERS
+)
 
 
 @click.group(name='tenant')
@@ -20,7 +19,7 @@ def describe(tenant_name=None):
     """
     Describes Tenant.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_get(tenant_name=tenant_name)
 
 
@@ -40,7 +39,7 @@ def activate(name, display_name, customer, cloud, read_only):
     """
     Activates Tenant.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_post(
         tenant_name=name,
         display_name=display_name,
@@ -58,7 +57,7 @@ def deactivate(tenant_name=None):
     """
     Deactivates Tenant.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_delete(tenant_name=tenant_name)
