+++ modular-service-admin-cli/group/tenant_region.py
@@ -1,9 +1,7 @@
 import click
 
-from modular_service_admin_cli.group import cli_response, ViewCommand
-from modular_service_admin_cli.service.constants import (PARAM_NAME,
-                                                         PARAM_PERMISSIONS,
-                                                         PARAM_ID)
+from group import cli_response, ViewCommand
+from service.constants import (PARAM_NAME, PARAM_PERMISSIONS, PARAM_ID)
 
 
 @click.group(name='region')
@@ -19,7 +17,7 @@ def describe(tenant_name):
     """
     Describes Tenant region.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_region_get(tenant_name=tenant_name)
 
 
@@ -33,7 +31,7 @@ def activate(tenant_name, region_name):
     """
     Activates region in tenant.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_region_post(
         tenant_name=tenant_name, region_name=region_name)
 
@@ -48,6 +46,6 @@ def deactivate(tenant_name, region_name):
     """
     Deactivates region in tenant.
     """
-    from modular_service_admin_cli.service.initializer import init_configuration
+    from service.initializer import init_configuration
     return init_configuration().tenant_region_delete(
         tenant_name=tenant_name, region_name=region_name)
