+++ modular-service-admin-cli/service/constants.py
