+++ modular-service-admin-cli/version.py
